Upgrading terraform to 0.10.2 and cloudfoundry provider to 0.9.1

+++ spec/tasks/terraform_apply_cloudfoundry/task_spec.rb
@@ -3,8 +3,9 @@ require 'yaml'
 require 'tmpdir'
 
 describe 'terraform_apply_cloudfoundry task' do
-  EXPECTED_TERRAFORM_VERSION='0.9.8'
-  EXPECTED_PROVIDER_CLOUDFOUNDRY_VERSION='v0.7.3'
+  EXPECTED_TERRAFORM_VERSION='0.10.2'
+  EXPECTED_PROVIDER_CLOUDFOUNDRY_VERSION='v0.9.1'
+  SKIP_TMP_FILE_CLEANUP=false
 
   context 'when pre-requisite are valid' do
     before(:context) do
@@ -59,8 +60,10 @@ describe 'terraform_apply_cloudfoundry task' do
     end
 
     after(:context) do
-      FileUtils.rm_rf @generated_files
-      FileUtils.rm_rf @spec_applied
+      unless SKIP_TMP_FILE_CLEANUP
+        FileUtils.rm_rf @generated_files
+        FileUtils.rm_rf @spec_applied
+      end
     end
 
     it 'applies to add only one resource' do
@@ -85,7 +88,7 @@ describe 'terraform_apply_cloudfoundry task' do
     end
 
     it 'matches files in generated-files output' do
-      expected_files = %w[. .. .gitkeep terraform.tfvars terraform.tfstate spec-only.txt].sort
+      expected_files = %w[. .. .gitkeep .terraform terraform.tfvars terraform.tfstate spec-only.txt].sort
       expect(Dir.entries(@generated_files).sort).to eq(expected_files)
     end
 
@@ -111,8 +114,10 @@ describe 'terraform_apply_cloudfoundry task' do
     end
 
     after(:context) do
-      FileUtils.rm_rf @generated_files
-      FileUtils.rm_rf @spec_applied
+      unless SKIP_TMP_FILE_CLEANUP
+        FileUtils.rm_rf @generated_files
+        FileUtils.rm_rf @spec_applied
+      end
     end
 
     it 'applies to add resources' do
@@ -163,9 +168,11 @@ describe 'terraform_apply_cloudfoundry task' do
     end
 
     after(:context) do
-      FileUtils.rm_rf @generated_files
-      FileUtils.rm_rf @spec_applied
-      FileUtils.rm_rf @terraform_tfvars
+      unless SKIP_TMP_FILE_CLEANUP
+        FileUtils.rm_rf @generated_files
+        FileUtils.rm_rf @spec_applied
+        FileUtils.rm_rf @terraform_tfvars
+      end
     end
 
     it 'applies to add only one resource' do
@@ -192,7 +199,7 @@ describe 'terraform_apply_cloudfoundry task' do
     end
 
     it 'matches files in generated-files output' do
-      expected_files = %w[. .. .gitkeep terraform.tfstate secrets.txt].sort
+      expected_files = %w[. .. .gitkeep .terraform terraform.tfstate secrets.txt].sort
       expect(Dir.entries(@generated_files).sort).to eq(expected_files)
     end
   end
